refactor(effects): split effect handlers and tests
