--- shadow_bout/effect_handlers.py
@@ -1,1778 +1,0 @@
-import random
-from dataclasses import replace
-from typing import Callable
-
-from shadow_bout.effect_utils import (
-    add_banned_card_id,
-    clear_forced_card_id,
-    get_opponent_side,
-    get_player_state,
-    is_immune_to_opponent_effect,
-    set_battle_janken_override,
-    set_must_reveal_played_card,
-    set_must_reveal_played_card_rounds,
-    update_player,
-)
-from shadow_bout.models import (
-    Card,
-    GameState,
-    Janken,
-    PendingEffectContext,
-    PersistentPointEffect,
-    Phase,
-    PointMatchEffect,
-    Side,
-)
-
-JANKEN_NAMES = {
-    Janken.ROCK: "グー",
-    Janken.SCISSORS: "チョキ",
-    Janken.PAPER: "パー",
-    Janken.WILDCARD: "ワイルド",
-}
-
-EffectHandler = Callable[[GameState, Side, Card], GameState]
-_registry: dict[str, EffectHandler] = {}
-
-
-def register(effect_type: str):
-    def decorator(fn: EffectHandler):
-        _registry[effect_type] = fn
-        return fn
-
-    return decorator
-
-
-def get_effect_handler(effect_type: str) -> EffectHandler | None:
-    return _registry.get(effect_type)
-
-
-def _finish_interactive_effect(state: GameState, log_msg: str) -> GameState:
-    return replace(
-        state,
-        phase=Phase.EFFECT_RESOLUTION,
-        pending_effect_context=None,
-        battle_log=state.battle_log + [log_msg],
-    )
-
-
-def _find_card(cards: list[Card], card_id: str | None) -> Card | None:
-    if card_id is None:
-        return None
-    return next((card for card in cards if card.id == card_id), None)
-
-
-def _find_card_by_id(state: GameState, side: Side, card_id: str | None) -> Card | None:
-    p_state = get_player_state(state, side)
-    card_lists = [
-        p_state.hand,
-        p_state.deck,
-        p_state.discard,
-        p_state.won_cards,
-        p_state.draw_stock,
-    ]
-    if state.current_battle is not None:
-        card_lists.append(
-            [
-                state.current_battle.player_card,
-                state.current_battle.npc_card,
-            ]
-        )
-    return next(
-        (card for cards in card_lists for card in cards if card.id == card_id), None
-    )
-
-
-def _remove_first_card_by_id(cards: list[Card], card_id: str) -> list[Card]:
-    removed = False
-    remaining: list[Card] = []
-    for card in cards:
-        if not removed and card.id == card_id:
-            removed = True
-            continue
-        remaining.append(card)
-    return remaining
-
-
-def _parse_card_ids(choice: str | None) -> list[str]:
-    if not choice:
-        return []
-    return [card_id for card_id in choice.split(",") if card_id]
-
-
-def _immune_blocked_state(state: GameState, card: Card) -> GameState:
-    return replace(
-        state,
-        battle_log=state.battle_log
-        + [f"{card.name}の効果発動: 相手は戦具効果を受けないため不発"],
-    )
-
-
-def _opponent_is_immune(state: GameState, side: Side) -> bool:
-    return is_immune_to_opponent_effect(state, side, get_opponent_side(side))
-
-
-def _resume_choose(state: GameState, side: Side, choice: str | None) -> GameState:
-    ctx = state.pending_effect_context
-    if ctx is None:
-        return state
-    p_state = get_player_state(state, side)
-    variant = ctx.payload.get("choose_variant")
-    if variant in ("yuriko_choose", "yuriko_return_cards"):
-        if ctx.step == 1:
-            return_count = int(ctx.payload.get("return_count", 0))
-            selected_ids = _parse_card_ids(choice)
-            returned = [
-                card
-                for card_id in selected_ids
-                if (card := _find_card(p_state.hand, card_id))
-            ][:return_count]
-
-            if len(returned) < return_count:
-                returned_ids = {card.id for card in returned}
-                returned.extend(
-                    card for card in p_state.hand if card.id not in returned_ids
-                )
-                returned = returned[:return_count]
-
-            returned_ids = {card.id for card in returned}
-            new_hand = [card for card in p_state.hand if card.id not in returned_ids]
-            new_deck = p_state.deck + returned
-            state = update_player(state, side, hand=new_hand, deck=new_deck)
-            return _finish_interactive_effect(
-                state,
-                "-> 百合子の効果: 山札から2枚引き、2枚を山札の下へ戻した",
-            )
-
-        if choice == "draw_cards":
-            drawn = p_state.deck[:2]
-            new_deck = p_state.deck[2:]
-            new_hand = p_state.hand + drawn
-            state = update_player(state, side, hand=new_hand, deck=new_deck)
-            return_count = min(2, len(new_hand))
-            if return_count == 0:
-                return _finish_interactive_effect(
-                    state,
-                    "-> 百合子の効果: 山札からカードを引けず、戻す手札もなかった",
-                )
-
-            next_ctx = replace(
-                ctx,
-                step=1,
-                payload={
-                    "choose_variant": "yuriko_return_cards",
-                    "return_count": return_count,
-                    "drawn_ids": [c.id for c in drawn],
-                },
-            )
-            return replace(
-                state,
-                phase=Phase.INTERACTIVE_EFFECT,
-                pending_effect_context=next_ctx,
-                battle_log=state.battle_log
-                + [
-                    f"-> 百合子の効果: 山札から{len(drawn)}枚引いた。戻す手札を選択中..."
-                ],
-            )
-
-        state = update_player(state, side, point_modifier=p_state.point_modifier + 3)
-        return _finish_interactive_effect(state, "-> 百合子の効果: ポイント+3")
-
-    if variant == "karen_choose":
-        if choice != "activate":
-            return _finish_interactive_effect(state, "-> 可憐の効果: 発動しない")
-        if _opponent_is_immune(state, side):
-            return _finish_interactive_effect(
-                state, "-> 可憐の効果: 相手は戦具効果を受けないため不発"
-            )
-        opp_side = get_opponent_side(side)
-        state = set_must_reveal_played_card_rounds(state, opp_side, rounds=2)
-        return _finish_interactive_effect(
-            state, "-> 可憐の効果: 相手は2ラウンドの間、出し札を公開"
-        )
-
-    return _finish_interactive_effect(state, "-> 選択効果: 未対応カード")
-
-
-def _resume_copy_hand(state: GameState, side: Side, choice: str | None) -> GameState:
-    p_state = get_player_state(state, side)
-    target = _find_card(p_state.hand, choice) or (
-        p_state.hand[0] if p_state.hand else None
-    )
-    if target is None or not target.effect or target.effect.type.value not in _registry:
-        return _finish_interactive_effect(state, "-> 杏奈の効果: 発動できる手札なし")
-
-    state = _finish_interactive_effect(
-        state, f"-> 杏奈の効果: {target.name}の効果を発動"
-    )
-    return replace(state, effect_queue=[(side, target)] + state.effect_queue)
-
-
-def _resume_search_and_swap(
-    state: GameState, side: Side, choice: str | None
-) -> GameState:
-    if not choice or ":" not in choice:
-        return _finish_interactive_effect(state, "-> 千鶴の効果: 交換しない")
-
-    hand_id, deck_id = choice.split(":", 1)
-    p_state = get_player_state(state, side)
-    hand_card = _find_card(p_state.hand, hand_id)
-    deck_card = _find_card(p_state.deck, deck_id)
-    if hand_card is None or deck_card is None:
-        return _finish_interactive_effect(state, "-> 千鶴の効果: 交換しない")
-
-    new_hand = [deck_card if card.id == hand_card.id else card for card in p_state.hand]
-    new_deck = [hand_card if card.id == deck_card.id else card for card in p_state.deck]
-    state = update_player(state, side, hand=new_hand, deck=new_deck)
-    return _finish_interactive_effect(
-        state, f"-> 千鶴の効果: {hand_card.name}と{deck_card.name}を交換"
-    )
-
-
-def _resume_swap(state: GameState, side: Side, choice: str | None) -> GameState:
-    p_state = get_player_state(state, side)
-    target = _find_card(p_state.hand, choice)
-    if target is None:
-        return _finish_interactive_effect(state, "-> 真の効果: 入れ替えない")
-
-    res = state.current_battle
-    if side == Side.PLAYER:
-        old_card = res.player_card
-        new_res = replace(res, player_card=target)
-    else:
-        old_card = res.npc_card
-        new_res = replace(res, npc_card=target)
-
-    new_hand = [card for card in p_state.hand if card.id != target.id] + [old_card]
-    state = update_player(state, side, hand=new_hand)
-    state = replace(state, current_battle=new_res)
-    return _finish_interactive_effect(
-        state, f"-> 真の効果: {old_card.name}と{target.name}を入れ替え"
-    )
-
-
-def _resume_tutor_play(state: GameState, side: Side, choice: str | None) -> GameState:
-    if choice in (None, "", "skip"):
-        return _finish_interactive_effect(state, "-> 麗花の効果: 発動しない")
-
-    p_state = get_player_state(state, side)
-    res = state.current_battle
-    if res is None:
-        return _finish_interactive_effect(
-            state, "-> 麗花の効果: 場のカードがないため不発"
-        )
-
-    old_card = res.player_card if side == Side.PLAYER else res.npc_card
-    target = _find_card(p_state.deck, choice)
-
-    if choice == old_card.id:
-        new_deck = list(p_state.deck)
-        random.shuffle(new_deck)
-        state = update_player(state, side, deck=new_deck)
-        return _finish_interactive_effect(
-            state, f"-> 麗花の効果: {old_card.name}を場に出し直し、山札を切った"
-        )
-
-    if target is None:
-        return _finish_interactive_effect(state, "-> 麗花の効果: 発動しない")
-
-    if side == Side.PLAYER:
-        new_res = replace(res, player_card=target)
-    else:
-        new_res = replace(res, npc_card=target)
-
-    new_deck = [card for card in p_state.deck if card.id != target.id] + [old_card]
-    random.shuffle(new_deck)
-    state = update_player(state, side, deck=new_deck)
-    state = replace(state, current_battle=new_res)
-    return _finish_interactive_effect(
-        state, f"-> 麗花の効果: {old_card.name}を山札に戻し、{target.name}を場に出した"
-    )
-
-
-def _resume_swap_opponent(
-    state: GameState, side: Side, choice: str | None
-) -> GameState:
-    ctx = state.pending_effect_context
-    opp_side = get_opponent_side(side)
-    source_card = _find_card_by_id(state, side, ctx.card_id if ctx else None)
-    if source_card and is_immune_to_opponent_effect(state, side, opp_side):
-        return _finish_interactive_effect(
-            state, f"-> {source_card.name}の効果: 相手は戦具効果を受けないため不発"
-        )
-
-    opp_state = get_player_state(state, opp_side)
-    if not opp_state.hand:
-        return _finish_interactive_effect(
-            state, "-> 可奈の効果: 相手の手札がないため入れ替えられない"
-        )
-    if ctx and ctx.step == 0:
-        target = random.choice(opp_state.hand)
-        next_ctx = replace(ctx, step=1, payload={**ctx.payload, "target_id": target.id})
-        return replace(
-            state,
-            pending_effect_context=next_ctx,
-            battle_log=state.battle_log
-            + [f"-> 可奈の効果: 相手手札から{target.name}を確認した"],
-        )
-
-    target = _find_card(opp_state.hand, ctx.payload.get("target_id") if ctx else None)
-    if target is None:
-        return _finish_interactive_effect(
-            state, "-> 可奈の効果: 確認したカードが手札にないため入れ替えない"
-        )
-    if choice != "swap":
-        return _finish_interactive_effect(state, "-> 可奈の効果: 入れ替えない")
-    res = state.current_battle
-    if opp_side == Side.PLAYER:
-        old_card = res.player_card
-        new_res = replace(res, player_card=target)
-    else:
-        old_card = res.npc_card
-        new_res = replace(res, npc_card=target)
-
-    new_opp_hand = [card for card in opp_state.hand if card.id != target.id] + [
-        old_card
-    ]
-    state = update_player(state, opp_side, hand=new_opp_hand)
-    state = replace(state, current_battle=new_res)
-    return _finish_interactive_effect(
-        state,
-        f"-> 可奈の効果: 相手手札を確認し、{old_card.name}と{target.name}を入れ替え",
-    )
-
-
-def _resume_salvage(state: GameState, side: Side, choice: str | None) -> GameState:
-    p_state = get_player_state(state, side)
-    if not p_state.discard:
-        return _finish_interactive_effect(state, "-> 風花の効果: 回収できる捨札がない")
-
-    target = _find_card(p_state.discard, choice)
-    if target is None:
-        return _finish_interactive_effect(state, "-> 風花の効果: 回収しない")
-
-    ctx = state.pending_effect_context
-    source_card = _find_card_by_id(state, side, ctx.card_id if ctx else None)
-    penalty = int(
-        source_card.effect.value if source_card and source_card.effect else -3
-    )
-    new_discard = [card for card in p_state.discard if card.id != target.id]
-    state = update_player(
-        state,
-        side,
-        hand=p_state.hand + [target],
-        discard=new_discard,
-        point_modifier=p_state.point_modifier + penalty,
-    )
-    return _finish_interactive_effect(
-        state, f"-> 風花の効果: {target.name}を回収し、ポイント{penalty}"
-    )
-
-
-def _resume_reorder(state: GameState, side: Side, choice: str | None) -> GameState:
-    p_state = get_player_state(state, side)
-    if not p_state.deck:
-        return _finish_interactive_effect(state, "-> 茜の効果: 並び替える山札がない")
-
-    selected_ids = _parse_card_ids(choice)
-    ordered: list[Card] = []
-    used_ids: set[str] = set()
-    for card_id in selected_ids:
-        card = _find_card(p_state.deck, card_id)
-        if card is None or card.id in used_ids:
-            continue
-        ordered.append(card)
-        used_ids.add(card.id)
-
-    ordered.extend(card for card in p_state.deck if card.id not in used_ids)
-    state = update_player(state, side, deck=ordered)
-    return _finish_interactive_effect(state, "-> 茜の効果: 山札の順番を入れ替えた")
-
-
-def _resume_choose_multiple(
-    state: GameState, side: Side, choice: str | None
-) -> GameState:
-    p_state = get_player_state(state, side)
-    if not choice:
-        return replace(
-            state,
-            battle_log=state.battle_log
-            + ["-> 海美の効果: 効果を1つ以上選択してください"],
-        )
-
-    selected = set(_parse_card_ids(choice))
-
-    can_discard = bool(p_state.hand)
-    can_draw = bool(p_state.deck)
-
-    do_discard = "discard_buff" in selected and can_discard
-    do_draw = "draw_debuff" in selected and can_draw
-
-    if not do_discard and not do_draw:
-        return _finish_interactive_effect(state, "-> 海美の効果: 発動可能な効果がない")
-
-    logs: list[str] = []
-
-    if do_discard:
-        discarded = p_state.hand[0]
-        p_state = replace(
-            p_state,
-            hand=p_state.hand[1:],
-            discard=p_state.discard + [discarded],
-            point_modifier=p_state.point_modifier + 5,
-        )
-        logs.append(f"{discarded.name}を捨札にしてポイント+5")
-
-    if do_draw:
-        drawn = p_state.deck[0]
-        p_state = replace(
-            p_state,
-            hand=p_state.hand + [drawn],
-            deck=p_state.deck[1:],
-            point_modifier=p_state.point_modifier - 2,
-        )
-        logs.append(f"{drawn.name}を1枚引いてポイント-2")
-
-    state = update_player(
-        state,
-        side,
-        hand=p_state.hand,
-        deck=p_state.deck,
-        discard=p_state.discard,
-        point_modifier=p_state.point_modifier,
-    )
-    return _finish_interactive_effect(state, f"-> 海美の効果: {'、'.join(logs)}")
-
-
-def _resume_debuff_counterable(
-    state: GameState, side: Side, choice: str | None
-) -> GameState:
-    ctx = state.pending_effect_context
-    source_side = get_opponent_side(side)
-    source_card = _find_card_by_id(state, source_side, ctx.card_id if ctx else None)
-    source_name = source_card.name if source_card else "志保"
-    debuff = int(
-        source_card.effect.value
-        if source_card and source_card.effect
-        else (ctx.payload.get("debuff", -5) if ctx else -5)
-    )
-
-    p_state = get_player_state(state, side)
-    discard_target = None
-    if choice == "counter":
-        discard_target = p_state.hand[0] if p_state.hand else None
-    elif choice not in (None, "", "skip", "no_counter"):
-        discard_target = _find_card(p_state.hand, choice)
-
-    if discard_target is None:
-        state = update_player(
-            state,
-            side,
-            point_modifier=p_state.point_modifier + debuff,
-        )
-        return _finish_interactive_effect(
-            state,
-            f"-> {source_name}の効果: 相手が無効化せず、相手のポイント{debuff:+d}",
-        )
-
-    state = update_player(
-        state,
-        side,
-        hand=[card for card in p_state.hand if card.id != discard_target.id],
-        discard=p_state.discard + [discard_target],
-    )
-    return _finish_interactive_effect(
-        state,
-        f"-> {source_name}の効果: 相手は{discard_target.name}を捨札にして無効化",
-    )
-
-
-def _coerce_declared_janken(choice: str | None) -> Janken:
-    if choice is None:
-        return Janken.ROCK
-    try:
-        janken = Janken(choice)
-    except ValueError:
-        return Janken.ROCK
-    if janken not in (Janken.ROCK, Janken.SCISSORS, Janken.PAPER):
-        return Janken.ROCK
-    return janken
-
-
-def _resume_special(state: GameState, side: Side, choice: str | None) -> GameState:
-    ctx = state.pending_effect_context
-    source_card = _find_card_by_id(state, side, ctx.card_id if ctx else None)
-    source_name = source_card.name if source_card else "美希"
-    opp_side = get_opponent_side(side)
-    opp_state = get_player_state(state, opp_side)
-
-    if ctx and ctx.step == 1:
-        target = _find_card(opp_state.won_cards, choice) or (
-            opp_state.won_cards[0] if opp_state.won_cards else None
-        )
-        if target is None:
-            return _finish_interactive_effect(
-                state, f"-> {source_name}の効果: 相手の勝ち札がないため奪取できない"
-            )
-
-        own_state = get_player_state(state, side)
-        state = update_player(state, side, deck=own_state.deck + [target])
-        state = update_player(
-            state,
-            opp_side,
-            won_cards=_remove_first_card_by_id(opp_state.won_cards, target.id),
-        )
-        return _finish_interactive_effect(
-            state,
-            f"-> {source_name}の効果: {target.name}を相手の勝ち札から奪い、自分の山札の下へ置いた",
-        )
-
-    declared = _coerce_declared_janken(choice)
-    declared_name = JANKEN_NAMES[declared]
-    if not opp_state.hand:
-        return _finish_interactive_effect(
-            state,
-            f"-> {source_name}の効果: {declared_name}を宣言したが、相手の手札がないため確認できない",
-        )
-
-    revealed = random.choice(opp_state.hand)
-    revealed_name = JANKEN_NAMES.get(revealed.janken, revealed.janken.value)
-    log = (
-        f"-> {source_name}の効果: {declared_name}を宣言し、"
-        f"相手手札から{revealed.name}({revealed_name})を確認"
-    )
-    if revealed.janken != declared:
-        return _finish_interactive_effect(state, f"{log}。一致しないため奪取なし")
-
-    if not opp_state.won_cards:
-        return _finish_interactive_effect(
-            state, f"{log}。一致したが相手の勝ち札がないため奪取できない"
-        )
-
-    next_ctx = replace(
-        ctx,
-        step=1,
-        payload={
-            **(ctx.payload if ctx else {}),
-            "declared_janken": declared.value,
-            "revealed_card_id": revealed.id,
-        },
-    )
-    return replace(
-        state,
-        phase=Phase.INTERACTIVE_EFFECT,
-        pending_effect_context=next_ctx,
-        battle_log=state.battle_log + [f"{log}。奪取する勝ち札を選択中..."],
-    )
-
-
-def _resume_removal(state: GameState, side: Side, choice: str | None) -> GameState:
-    if choice not in (None, "activate", "yes", "true"):
-        return _finish_interactive_effect(state, "-> ジュリアの効果: 発動しない")
-
-    ctx = state.pending_effect_context
-    source_card = _find_card_by_id(state, side, ctx.card_id if ctx else None)
-    if source_card and _opponent_is_immune(state, side):
-        return _finish_interactive_effect(
-            state, f"-> {source_card.name}の効果: 相手は戦具効果を受けないため不発"
-        )
-
-    res = state.current_battle
-    if side == Side.PLAYER:
-        own_state = state.player
-        opp_state = state.npc
-        state = replace(
-            state,
-            player=replace(own_state, deck=own_state.deck + [res.player_card]),
-            npc=replace(opp_state, discard=opp_state.discard + [res.npc_card]),
-        )
-    else:
-        own_state = state.npc
-        opp_state = state.player
-        state = replace(
-            state,
-            npc=replace(own_state, deck=own_state.deck + [res.npc_card]),
-            player=replace(opp_state, discard=opp_state.discard + [res.player_card]),
-        )
-
-    state = replace(state, removal_activated=True)
-    return _finish_interactive_effect(state, "-> ジュリアの効果: 勝敗判定をスキップ")
-
-
-def _resume_set_point_match(
-    state: GameState, side: Side, choice: str | None
-) -> GameState:
-    ctx = state.pending_effect_context
-    source_card = _find_card_by_id(state, side, ctx.card_id if ctx else None)
-    source_name = source_card.name if source_card else "歩"
-
-    if choice != "activate":
-        return _finish_interactive_effect(state, f"-> {source_name}の効果: 発動しない")
-
-    if _opponent_is_immune(state, side):
-        return _finish_interactive_effect(
-            state, f"-> {source_name}の効果: 相手は戦具効果を受けないため不発"
-        )
-
-    effect = PointMatchEffect(
-        source_side=side,
-        target_side=get_opponent_side(side),
-    )
-    state = replace(
-        state,
-        point_match_effects=state.point_match_effects + (effect,),
-    )
-    return _finish_interactive_effect(
-        state, f"-> {source_name}の効果: 相手のポイントをこのカードと同じにする"
-    )
-
-
-def resume_pending_effect(state: GameState, choice: str | None = None) -> GameState:
-    ctx = state.pending_effect_context
-    if not ctx:
-        return state
-
-    side = ctx.side
-    if ctx.effect == "choose":
-        return _resume_choose(state, side, choice)
-    if ctx.effect == "copy_hand":
-        return _resume_copy_hand(state, side, choice)
-    if ctx.effect == "search_and_swap":
-        return _resume_search_and_swap(state, side, choice)
-    if ctx.effect == "swap":
-        return _resume_swap(state, side, choice)
-    if ctx.effect == "tutor_play":
-        return _resume_tutor_play(state, side, choice)
-    if ctx.effect == "swap_opponent":
-        return _resume_swap_opponent(state, side, choice)
-    if ctx.effect == "removal":
-        return _resume_removal(state, side, choice)
-    if ctx.effect == "set_point_match":
-        return _resume_set_point_match(state, side, choice)
-    if ctx.effect == "salvage":
-        return _resume_salvage(state, side, choice)
-    if ctx.effect == "reorder":
-        return _resume_reorder(state, side, choice)
-    if ctx.effect == "choose_multiple":
-        return _resume_choose_multiple(state, side, choice)
-    if ctx.effect == "debuff_counterable":
-        return _resume_debuff_counterable(state, side, choice)
-    if ctx.effect == "special":
-        return _resume_special(state, side, choice)
-
-    return _finish_interactive_effect(state, "-> 選択完了")
-
-
-@register("null")
-def effect_null(state: GameState, side: Side, card: Card) -> GameState:
-    return state
-
-
-@register("immune")
-def effect_immune(state: GameState, side: Side, card: Card) -> GameState:
-    return replace(
-        state,
-        battle_log=state.battle_log
-        + [f"{card.name}の効果発動: 相手の戦具効果を受けない"],
-    )
-
-
-@register("buff")
-def effect_buff(state: GameState, side: Side, card: Card) -> GameState:
-    p_state = get_player_state(state, side)
-    new_mod = p_state.point_modifier + int(card.effect.value or 0)
-    state = update_player(state, side, point_modifier=new_mod)
-    return replace(
-        state,
-        battle_log=state.battle_log
-        + [f"{card.name}の効果発動: ポイント+{card.effect.value}"],
-    )
-
-
-@register("buff_and_peek")
-def effect_buff_and_peek(state: GameState, side: Side, card: Card) -> GameState:
-    p_state = get_player_state(state, side)
-    bonus = int(card.effect.value or 0)
-    state = update_player(state, side, point_modifier=p_state.point_modifier + bonus)
-
-    updated_state = get_player_state(state, side)
-    if not updated_state.deck:
-        return replace(
-            state,
-            battle_log=state.battle_log
-            + [f"{card.name}の効果発動: ポイント+{bonus}、山札が空のため確認できない"],
-        )
-
-    top_card = updated_state.deck[0]
-    return replace(
-        state,
-        battle_log=state.battle_log
-        + [f"{card.name}の効果発動: ポイント+{bonus}、山札の一番上は{top_card.name}"],
-    )
-
-
-@register("draw")
-def effect_draw(state: GameState, side: Side, card: Card) -> GameState:
-    if card.id in ("card_01", "c1"):
-        own_state = get_player_state(state, side)
-        own_merged_deck = own_state.deck + own_state.hand
-        random.shuffle(own_merged_deck)
-        own_draw_count = int(card.effect.value or 0)
-        own_drawn = own_merged_deck[:own_draw_count]
-        state = update_player(
-            state,
-            side,
-            hand=own_drawn,
-            deck=own_merged_deck[len(own_drawn) :],
-        )
-
-        opp_side = get_opponent_side(side)
-        if _opponent_is_immune(state, side):
-            return replace(
-                state,
-                battle_log=state.battle_log
-                + [
-                    f"{card.name}の効果発動: 自分は{len(own_drawn)}枚ドロー。相手は戦具効果を受けないため対象外"
-                ],
-            )
-
-        opp_state = get_player_state(state, opp_side)
-        opp_merged_deck = opp_state.deck + opp_state.hand
-        random.shuffle(opp_merged_deck)
-        opp_draw_count = int(card.effect.value or 0)
-        opp_drawn = opp_merged_deck[:opp_draw_count]
-        state = update_player(
-            state,
-            opp_side,
-            hand=opp_drawn,
-            deck=opp_merged_deck[len(opp_drawn) :],
-        )
-
-        return replace(
-            state,
-            battle_log=state.battle_log
-            + [
-                f"{card.name}の効果発動: お互いの手札を山札に戻して切り、自分は{len(own_drawn)}枚、相手は{len(opp_drawn)}枚ドロー"
-            ],
-        )
-
-    draw_count = int(card.effect.value or 0)
-    own_state = get_player_state(state, side)
-    own_drawn = own_state.deck[:draw_count]
-    state = update_player(
-        state,
-        side,
-        hand=own_state.hand + own_drawn,
-        deck=own_state.deck[len(own_drawn) :],
-    )
-
-    opp_side = get_opponent_side(side)
-    if _opponent_is_immune(state, side):
-        return replace(
-            state,
-            battle_log=state.battle_log
-            + [
-                f"{card.name}の効果発動: 自分は{len(own_drawn)}枚ドロー。相手は戦具効果を受けないため対象外"
-            ],
-        )
-
-    opp_state = get_player_state(state, opp_side)
-    opp_drawn = opp_state.deck[:draw_count]
-    state = update_player(
-        state,
-        opp_side,
-        hand=opp_state.hand + opp_drawn,
-        deck=opp_state.deck[len(opp_drawn) :],
-    )
-
-    return replace(
-        state,
-        battle_log=state.battle_log
-        + [
-            f"{card.name}の効果発動: 自分は{len(own_drawn)}枚、相手は{len(opp_drawn)}枚ドロー"
-        ],
-    )
-
-
-@register("draw_dynamic")
-def effect_draw_dynamic(state: GameState, side: Side, card: Card) -> GameState:
-    own_state = get_player_state(state, side)
-    moved_count = len(own_state.discard)
-
-    if moved_count == 0:
-        return replace(
-            state,
-            battle_log=state.battle_log + [f"{card.name}の効果発動: 捨札がなく不発"],
-        )
-
-    new_deck = own_state.deck + own_state.discard
-    random.shuffle(new_deck)
-    drawn = new_deck[:moved_count]
-    state = update_player(
-        state,
-        side,
-        hand=own_state.hand + drawn,
-        deck=new_deck[len(drawn) :],
-        discard=[],
-    )
-    return replace(
-        state,
-        battle_log=state.battle_log
-        + [
-            f"{card.name}の効果発動: 捨札{moved_count}枚を山札に戻して{len(drawn)}枚ドロー"
-        ],
-    )
-
-
-@register("steal_draw")
-def effect_steal_draw(state: GameState, side: Side, card: Card) -> GameState:
-    opp_side = get_opponent_side(side)
-    if _opponent_is_immune(state, side):
-        return _immune_blocked_state(state, card)
-
-    opp_state = get_player_state(state, opp_side)
-
-    if not opp_state.deck:
-        return replace(
-            state,
-            battle_log=state.battle_log
-            + [f"{card.name}の効果発動: 相手の山札が空のため不発"],
-        )
-
-    stolen = opp_state.deck[0]
-    state = update_player(
-        state,
-        opp_side,
-        deck=opp_state.deck[1:],
-    )
-
-    own_state = get_player_state(state, side)
-    state = update_player(
-        state,
-        side,
-        hand=own_state.hand + [stolen],
-    )
-
-    return replace(
-        state,
-        battle_log=state.battle_log
-        + [
-            f"{card.name}の効果発動: 相手の山札の上から{stolen.name}を奪って手札に加えた"
-        ],
-    )
-
-
-@register("steal_hand")
-def effect_steal_hand(state: GameState, side: Side, card: Card) -> GameState:
-    opp_side = get_opponent_side(side)
-    if _opponent_is_immune(state, side):
-        return _immune_blocked_state(state, card)
-
-    opp_state = get_player_state(state, opp_side)
-
-    if not opp_state.hand:
-        return replace(
-            state,
-            battle_log=state.battle_log
-            + [f"{card.name}の効果発動: 相手の手札が空のため不発"],
-        )
-
-    stolen_idx = random.randrange(len(opp_state.hand))
-    stolen = opp_state.hand[stolen_idx]
-    state = update_player(
-        state,
-        opp_side,
-        hand=[c for i, c in enumerate(opp_state.hand) if i != stolen_idx],
-    )
-
-    own_state = get_player_state(state, side)
-    state = update_player(
-        state,
-        side,
-        hand=own_state.hand + [stolen],
-    )
-
-    return replace(
-        state,
-        battle_log=state.battle_log
-        + [f"{card.name}の効果発動: 相手の手札から{stolen.name}を奪って手札に加えた"],
-    )
-
-
-@register("special")
-def effect_special(state: GameState, side: Side, card: Card) -> GameState:
-    if _opponent_is_immune(state, side):
-        return _immune_blocked_state(state, card)
-
-    ctx = PendingEffectContext(side=side, card_id=card.id, effect="special")
-    state = replace(
-        state, phase=Phase.INTERACTIVE_EFFECT, effect_step=0, pending_effect_context=ctx
-    )
-    return replace(
-        state,
-        battle_log=state.battle_log + [f"{card.name}の効果発動: 宣言待機中..."],
-    )
-
-
-@register("win_condition")
-def effect_win_condition(state: GameState, side: Side, card: Card) -> GameState:
-    own_state = get_player_state(state, side)
-    opp_side = get_opponent_side(side)
-    opp_state = get_player_state(state, opp_side)
-
-    own_top = own_state.deck[0] if own_state.deck else None
-    opp_top = opp_state.deck[0] if opp_state.deck else None
-
-    if own_top is not None:
-        state = update_player(state, side, deck=own_state.deck[1:] + [own_top])
-    if opp_top is not None:
-        state = update_player(state, opp_side, deck=opp_state.deck[1:] + [opp_top])
-
-    if own_top is None or opp_top is None:
-        return replace(
-            state,
-            battle_log=state.battle_log
-            + [f"{card.name}の効果発動: 山札が不足しているため不発"],
-        )
-
-    revealed_jankens = {own_top.janken, opp_top.janken}
-    reveal_log = (
-        f"{card.name}の効果発動: 山札の上から自分は{own_top.name}"
-        f"({JANKEN_NAMES.get(own_top.janken, own_top.janken.value)})、"
-        f"相手は{opp_top.name}"
-        f"({JANKEN_NAMES.get(opp_top.janken, opp_top.janken.value)})"
-        "を公開し山札の下へ戻した"
-    )
-
-    if revealed_jankens == {Janken.PAPER, Janken.SCISSORS}:
-        return replace(
-            state,
-            win_condition_winner=side,
-            battle_log=state.battle_log + [f"{reveal_log}。勝利条件成立"],
-        )
-
-    return replace(
-        state,
-        battle_log=state.battle_log + [f"{reveal_log}。勝利条件不成立"],
-    )
-
-
-@register("buff_dynamic")
-def effect_buff_dynamic(state: GameState, side: Side, card: Card) -> GameState:
-    p_state = get_player_state(state, side)
-    bonus = len(p_state.hand) * int(card.effect.value or 1)
-    new_mod = p_state.point_modifier + bonus
-    state = update_player(state, side, point_modifier=new_mod)
-    return replace(
-        state,
-        battle_log=state.battle_log
-        + [f"{card.name}の効果発動: 手札枚数x{card.effect.value} = +{bonus}"],
-    )
-
-
-@register("buff_next")
-def effect_buff_next(state: GameState, side: Side, card: Card) -> GameState:
-    p_state = get_player_state(state, side)
-    bonus = int(card.effect.value or 0)
-    state = update_player(
-        state,
-        side,
-        next_round_point_modifier=p_state.next_round_point_modifier + bonus,
-    )
-    return replace(
-        state,
-        battle_log=state.battle_log
-        + [f"{card.name}の効果発動: 次ラウンドのポイント{bonus:+d}"],
-    )
-
-
-@register("buff_snowball")
-def effect_buff_snowball(state: GameState, side: Side, card: Card) -> GameState:
-    p_state = get_player_state(state, side)
-    bonus = int(card.effect.value or 0)
-    state = update_player(
-        state,
-        side,
-        point_modifier=p_state.point_modifier + bonus,
-    )
-    return replace(
-        state,
-        pending_next_round_buff_on_win=(
-            state.pending_next_round_buff_on_win + ((side, bonus),)
-        ),
-        battle_log=state.battle_log
-        + [
-            f"{card.name}の効果発動: この勝負のポイント{bonus:+d}、"
-            f"勝利時に次ラウンドのポイント{bonus:+d}"
-        ],
-    )
-
-
-@register("change_janken")
-def effect_change_janken(state: GameState, side: Side, card: Card) -> GameState:
-    opp_side = get_opponent_side(side)
-    if _opponent_is_immune(state, side):
-        return _immune_blocked_state(state, card)
-
-    target_janken = Janken(card.effect.value)
-    target_name = JANKEN_NAMES[target_janken]
-
-    # card_09（律子）は「次の勝負」の相手の場/手札に適用する。
-    if card.id in ("card_09", "c9"):
-        state = update_player(
-            state,
-            opp_side,
-            next_round_janken_override=target_janken,
-        )
-        return replace(
-            state,
-            battle_log=state.battle_log
-            + [f"{card.name}の効果発動: 次ラウンドの相手の場と手札を{target_name}扱い"],
-        )
-
-    # card_28（亜利沙）は即時に現在の相手場カードだけへ適用し、効果後の再判定に反映する。
-    if card.id in ("card_28", "c28"):
-        state = set_battle_janken_override(state, opp_side, target_janken)
-        return replace(
-            state,
-            battle_log=state.battle_log
-            + [f"{card.name}の効果発動: 相手の場のカードを{target_name}扱い"],
-        )
-
-    return replace(
-        state,
-        battle_log=state.battle_log + [f"{card.name}の効果発動: 未対応カードID"],
-    )
-
-
-@register("buff_scaling")
-def effect_buff_scaling(state: GameState, side: Side, card: Card) -> GameState:
-    p_state = get_player_state(state, side)
-    bonus = state.round_number * 2
-    state = update_player(state, side, point_modifier=p_state.point_modifier + bonus)
-    return replace(
-        state,
-        battle_log=state.battle_log
-        + [f"{card.name}の効果発動: R{state.round_number}のためポイント+{bonus}"],
-    )
-
-
-@register("debuff")
-def effect_debuff(state: GameState, side: Side, card: Card) -> GameState:
-    opp_side = get_opponent_side(side)
-    if _opponent_is_immune(state, side):
-        return _immune_blocked_state(state, card)
-
-    opp_state = get_player_state(state, opp_side)
-
-    if card.id in ("card_04", "c4"):
-        current_battle = state.current_battle
-        opp_card = (
-            current_battle.npc_card
-            if opp_side == Side.NPC
-            else current_battle.player_card
-        )
-        conditional = (
-            opp_state.conditional_point_modifier_non_wildcard
-            if opp_card.janken != Janken.WILDCARD
-            else 0
-        )
-        current_point = opp_card.base_point + opp_state.point_modifier + conditional
-        halved_point = current_point // 2
-        new_modifier = halved_point - opp_card.base_point - conditional
-        state = update_player(state, opp_side, point_modifier=new_modifier)
-        return replace(
-            state,
-            battle_log=state.battle_log
-            + [f"{card.name}の効果発動: 相手のポイントを半分(切り捨て)にした"],
-        )
-
-    if card.id in ("card_17", "c17"):
-        debuff = len(opp_state.hand) * int(card.effect.value or 0)
-        state = update_player(
-            state, opp_side, point_modifier=opp_state.point_modifier + debuff
-        )
-        return replace(
-            state,
-            battle_log=state.battle_log
-            + [f"{card.name}の効果発動: 相手の手札枚数ぶんポイント{debuff:+d}"],
-        )
-
-    return state
-
-
-@register("debuff_counterable")
-def effect_debuff_counterable(state: GameState, side: Side, card: Card) -> GameState:
-    opp_side = get_opponent_side(side)
-    if _opponent_is_immune(state, side):
-        return _immune_blocked_state(state, card)
-
-    debuff = int(card.effect.value or 0)
-    opp_state = get_player_state(state, opp_side)
-    if not opp_state.hand:
-        state = update_player(
-            state,
-            opp_side,
-            point_modifier=opp_state.point_modifier + debuff,
-        )
-        return replace(
-            state,
-            battle_log=state.battle_log
-            + [f"{card.name}の効果発動: 相手の手札がないため相手のポイント{debuff:+d}"],
-        )
-
-    ctx = PendingEffectContext(
-        side=opp_side,
-        card_id=card.id,
-        effect="debuff_counterable",
-        payload={"debuff": debuff},
-    )
-    state = replace(
-        state, phase=Phase.INTERACTIVE_EFFECT, effect_step=0, pending_effect_context=ctx
-    )
-    return replace(
-        state,
-        battle_log=state.battle_log
-        + [f"{card.name}の効果発動: 相手の無効化選択待機中..."],
-    )
-
-
-@register("set_point")
-def effect_set_point(state: GameState, side: Side, card: Card) -> GameState:
-    opp_side = get_opponent_side(side)
-    if _opponent_is_immune(state, side):
-        return _immune_blocked_state(state, card)
-
-    opp_state = get_player_state(state, opp_side)
-
-    set_value = int(card.effect.value or 0)
-    current_battle = state.current_battle
-    opp_card = (
-        current_battle.npc_card if opp_side == Side.NPC else current_battle.player_card
-    )
-    conditional = (
-        opp_state.conditional_point_modifier_non_wildcard
-        if opp_card.janken != Janken.WILDCARD
-        else 0
-    )
-    new_modifier = set_value - opp_card.base_point - conditional
-    state = update_player(state, opp_side, point_modifier=new_modifier)
-    return replace(
-        state,
-        battle_log=state.battle_log
-        + [f"{card.name}の効果発動: 相手のポイントを{set_value}にした"],
-    )
-
-
-@register("set_point_match")
-def effect_set_point_match(state: GameState, side: Side, card: Card) -> GameState:
-    ctx = PendingEffectContext(side=side, card_id=card.id, effect="set_point_match")
-    state = replace(
-        state, phase=Phase.INTERACTIVE_EFFECT, effect_step=0, pending_effect_context=ctx
-    )
-    return replace(
-        state,
-        battle_log=state.battle_log + [f"{card.name}の効果発動: 発動選択待機中..."],
-    )
-
-
-@register("conditional_buff")
-def effect_conditional_buff(state: GameState, side: Side, card: Card) -> GameState:
-    p_state = get_player_state(state, side)
-    opp_side = get_opponent_side(side)
-    opp_state = get_player_state(state, opp_side)
-
-    own_won_total = sum(won.base_point for won in p_state.won_cards)
-    opp_won_total = sum(won.base_point for won in opp_state.won_cards)
-    bonus = int(card.effect.value or 0)
-
-    if opp_won_total > own_won_total:
-        state = update_player(
-            state,
-            side,
-            point_modifier=p_state.point_modifier + bonus,
-        )
-        return replace(
-            state,
-            battle_log=state.battle_log
-            + [
-                f"{card.name}の効果発動: 相手の勝ち札合計({opp_won_total}) > 自分({own_won_total})のためポイント+{bonus}"
-            ],
-        )
-
-    return replace(
-        state,
-        battle_log=state.battle_log
-        + [
-            f"{card.name}の効果発動: 相手の勝ち札合計({opp_won_total}) <= 自分({own_won_total})のため不発"
-        ],
-    )
-
-
-@register("conditional_negate_buff")
-def effect_conditional_negate_buff(
-    state: GameState, side: Side, card: Card
-) -> GameState:
-    opp_side = get_opponent_side(side)
-    opponent_is_immune = _opponent_is_immune(state, side)
-
-    current_battle = state.current_battle
-    if current_battle is None:
-        return replace(
-            state,
-            battle_log=state.battle_log
-            + [f"{card.name}の効果発動: 場のカードがないため不発"],
-        )
-
-    p_state = get_player_state(state, side)
-    opp_state = get_player_state(state, opp_side)
-    opp_card = (
-        current_battle.npc_card if opp_side == Side.NPC else current_battle.player_card
-    )
-    conditional = (
-        opp_state.conditional_point_modifier_non_wildcard
-        if opp_card.janken != Janken.WILDCARD
-        else 0
-    )
-    opponent_point = opp_card.base_point + opp_state.point_modifier + conditional
-
-    if opponent_point % 2 == 0:
-        return replace(
-            state,
-            battle_log=state.battle_log
-            + [
-                f"{card.name}の効果発動: 相手のポイント({opponent_point})が偶数のため不発"
-            ],
-        )
-
-    bonus = int(card.effect.value or 0)
-    if not opponent_is_immune:
-        state = update_player(state, opp_side, effect_negated=True)
-    state = update_player(state, side, point_modifier=p_state.point_modifier + bonus)
-    if opponent_is_immune:
-        return replace(
-            state,
-            battle_log=state.battle_log
-            + [
-                f"{card.name}の効果発動: 相手のポイント({opponent_point})が奇数だが、相手は戦具効果を受けないため無効化できず、ポイント+{bonus}"
-            ],
-        )
-
-    return replace(
-        state,
-        battle_log=state.battle_log
-        + [
-            f"{card.name}の効果発動: 相手のポイント({opponent_point})が奇数のため相手の戦具効果を無効化し、ポイント+{bonus}"
-        ],
-    )
-
-
-@register("conditional_debuff_next")
-def effect_conditional_debuff_next(
-    state: GameState, side: Side, card: Card
-) -> GameState:
-    opp_side = get_opponent_side(side)
-    if _opponent_is_immune(state, side):
-        return _immune_blocked_state(state, card)
-
-    opp_state = get_player_state(state, opp_side)
-    debuff = int(card.effect.value or 0)
-    state = update_player(
-        state,
-        opp_side,
-        next_round_conditional_point_modifier_non_wildcard=(
-            opp_state.next_round_conditional_point_modifier_non_wildcard + debuff
-        ),
-    )
-    return replace(
-        state,
-        battle_log=state.battle_log
-        + [f"{card.name}の効果発動: 相手の次ラウンド(バー以外)のポイント{debuff:+d}"],
-    )
-
-
-@register("conditional_debuff_draw")
-def effect_conditional_debuff_draw(
-    state: GameState, side: Side, card: Card
-) -> GameState:
-    opp_side = get_opponent_side(side)
-    current_battle = state.current_battle
-    if current_battle is None:
-        return replace(
-            state,
-            battle_log=state.battle_log
-            + [f"{card.name}の効果発動: 場のカードがないため不発"],
-        )
-
-    opp_state = get_player_state(state, opp_side)
-    opp_card = (
-        current_battle.npc_card if opp_side == Side.NPC else current_battle.player_card
-    )
-    conditional = (
-        opp_state.conditional_point_modifier_non_wildcard
-        if opp_card.janken != Janken.WILDCARD
-        else 0
-    )
-    opponent_point = opp_card.base_point + opp_state.point_modifier + conditional
-    debuff = int(card.effect.value or 0)
-    draw_count = 1
-    logs = [f"勝利時に{draw_count}枚ドローを予約"]
-
-    if opponent_point <= 15:
-        if _opponent_is_immune(state, side):
-            logs.append(
-                f"相手のポイント({opponent_point})は15以下だが、"
-                "相手は戦具効果を受けないためポイント減少は不発"
-            )
-        else:
-            state = update_player(
-                state,
-                opp_side,
-                point_modifier=opp_state.point_modifier + debuff,
-            )
-            logs.append(
-                f"相手のポイント({opponent_point})が15以下のため相手{debuff:+d}"
-            )
-    else:
-        logs.append(f"相手のポイント({opponent_point})が16以上のためポイント減少は不発")
-
-    return replace(
-        state,
-        pending_draw_on_win=state.pending_draw_on_win + ((side, draw_count),),
-        battle_log=state.battle_log + [f"{card.name}の効果発動: {'、'.join(logs)}"],
-    )
-
-
-@register("debuff_persistent")
-def effect_debuff_persistent(state: GameState, side: Side, card: Card) -> GameState:
-    opp_side = get_opponent_side(side)
-    if _opponent_is_immune(state, side):
-        return _immune_blocked_state(state, card)
-
-    opp_state = get_player_state(state, opp_side)
-    debuff = int(card.effect.value or 0)
-    state = update_player(
-        state, opp_side, point_modifier=opp_state.point_modifier + debuff
-    )
-    # 「このターンと次のターン」: 現ターンは即時適用し、次ターン分を継続効果として保持
-    updated_opp_state = get_player_state(state, opp_side)
-    state = update_player(
-        state,
-        opp_side,
-        persistent_point_effects=updated_opp_state.persistent_point_effects
-        + (PersistentPointEffect(value=debuff, remaining_turns=1),),
-    )
-    return replace(
-        state,
-        battle_log=state.battle_log
-        + [f"{card.name}の効果発動: 相手のポイントをこのターンと次ターン{debuff:+d}"],
-    )
-
-
-@register("debuff_conditional")
-def effect_debuff_conditional(state: GameState, side: Side, card: Card) -> GameState:
-    opp_side = get_opponent_side(side)
-    if _opponent_is_immune(state, side):
-        return _immune_blocked_state(state, card)
-
-    opp_state = get_player_state(state, opp_side)
-    debuff = int(card.effect.value or 0)
-
-    if card.id == "card_18":
-        return replace(
-            state,
-            battle_log=state.battle_log
-            + [
-                f"{card.name}の効果発動: この勝負で敗北した場合に次ラウンド(バー以外)ポイント{debuff:+d}"
-            ],
-            pending_conditional_debuff_on_loss=(
-                state.pending_conditional_debuff_on_loss + ((side, debuff),)
-            ),
-        )
-
-    if card.id == "card_30":
-        if state.round_number >= 3:
-            state = update_player(
-                state,
-                opp_side,
-                next_round_conditional_point_modifier_non_wildcard=(
-                    opp_state.next_round_conditional_point_modifier_non_wildcard
-                    + debuff
-                ),
-            )
-            return replace(
-                state,
-                battle_log=state.battle_log
-                + [
-                    f"{card.name}の効果発動: 3戦目以降のため相手の次ラウンド(バー以外)ポイント{debuff:+d}"
-                ],
-            )
-
-        return replace(
-            state,
-            battle_log=state.battle_log
-            + [f"{card.name}の効果発動: 3戦目未満のため不発(R{state.round_number})"],
-        )
-
-    return replace(
-        state,
-        battle_log=state.battle_log
-        + [f"{card.name}の効果発動: 未対応の条件カードID({card.id})"],
-    )
-
-
-@register("negate")
-def effect_negate(state: GameState, side: Side, card: Card) -> GameState:
-    opp_side = get_opponent_side(side)
-    if _opponent_is_immune(state, side):
-        return _immune_blocked_state(state, card)
-
-    state = update_player(state, opp_side, effect_negated=True)
-    return replace(
-        state,
-        battle_log=state.battle_log
-        + [f"{card.name}の効果発動: 相手の戦具効果を無効化"],
-    )
-
-
-@register("force_play")
-def effect_force_play(state: GameState, side: Side, card: Card) -> GameState:
-    opp_side = get_opponent_side(side)
-    if _opponent_is_immune(state, side):
-        return _immune_blocked_state(state, card)
-
-    opp_state = get_player_state(state, opp_side)
-    if not opp_state.hand:
-        return replace(
-            state,
-            battle_log=state.battle_log
-            + [f"{card.name}の効果発動: 相手の手札がないため不発"],
-        )
-
-    target = random.choice(opp_state.hand)
-    state = update_player(state, opp_side, forced_card_id=target.id)
-    return replace(
-        state,
-        battle_log=state.battle_log
-        + [f"{card.name}の効果発動: 相手は次ラウンドで{target.name}を強制プレイ"],
-    )
-
-
-@register("ban")
-def effect_ban(state: GameState, side: Side, card: Card) -> GameState:
-    opp_side = get_opponent_side(side)
-    if _opponent_is_immune(state, side):
-        return _immune_blocked_state(state, card)
-
-    opp_state = get_player_state(state, opp_side)
-    if not opp_state.hand:
-        return replace(
-            state,
-            battle_log=state.battle_log
-            + [f"{card.name}の効果発動: 相手の手札がないため不発"],
-        )
-
-    target = random.choice(opp_state.hand)
-    state = add_banned_card_id(state, opp_side, target.id)
-
-    if opp_state.forced_card_id == target.id:
-        state = clear_forced_card_id(state, opp_side)
-        return replace(
-            state,
-            battle_log=state.battle_log
-            + [
-                f"{card.name}の効果発動: 相手の{target.name}を使用禁止（強制プレイは解除）"
-            ],
-        )
-
-    return replace(
-        state,
-        battle_log=state.battle_log
-        + [f"{card.name}の効果発動: 相手の{target.name}を使用禁止"],
-    )
-
-
-@register("reveal")
-def effect_reveal(state: GameState, side: Side, card: Card) -> GameState:
-    opp_side = get_opponent_side(side)
-    if _opponent_is_immune(state, side):
-        return _immune_blocked_state(state, card)
-
-    opp_state = get_player_state(state, opp_side)
-    if not opp_state.hand:
-        return state
-
-    target = random.choice(opp_state.hand)
-    new_revealed = opp_state.revealed_card_ids | {target.id}
-    state = update_player(state, opp_side, revealed_card_ids=new_revealed)
-    return replace(
-        state,
-        battle_log=state.battle_log
-        + [f"{card.name}の効果発動: 相手の手札を1枚永続公開"],
-    )
-
-
-@register("reveal_all")
-def effect_reveal_all(state: GameState, side: Side, card: Card) -> GameState:
-    opp_side = get_opponent_side(side)
-    if _opponent_is_immune(state, side):
-        return _immune_blocked_state(state, card)
-
-    opp_state = get_player_state(state, opp_side)
-    new_revealed = opp_state.revealed_card_ids | {
-        hand_card.id for hand_card in opp_state.hand
-    }
-    state = update_player(state, opp_side, revealed_card_ids=new_revealed)
-
-    state = replace(
-        state,
-        revealed_this_round=list(opp_state.hand),
-        revealed_this_round_side=opp_side,
-    )
-    return replace(
-        state,
-        battle_log=state.battle_log + [f"{card.name}の効果発動: 相手の手札を全て公開"],
-    )
-
-
-@register("restart")
-def effect_restart(state: GameState, side: Side, card: Card) -> GameState:
-    if _opponent_is_immune(state, side):
-        return _immune_blocked_state(state, card)
-
-    if state.last_restart_round == state.round_number - 1:
-        return replace(
-            state,
-            battle_log=state.battle_log
-            + [f"{card.name}の効果発動: 連続使用不可のため不発"],
-        )
-
-    state = replace(
-        state,
-        battle_log=state.battle_log + [f"{card.name}の効果発動: この勝負をやり直す"],
-    )
-
-    res = state.current_battle
-    p_card = res.player_card
-    n_card = res.npc_card
-
-    new_p_hand = state.player.hand + [p_card]
-    new_n_hand = state.npc.hand + [n_card]
-
-    state = update_player(state, Side.PLAYER, hand=new_p_hand)
-    state = update_player(state, Side.NPC, hand=new_n_hand)
-
-    state = update_player(state, Side.PLAYER, point_modifier=0, effect_negated=False)
-    state = update_player(state, Side.NPC, point_modifier=0, effect_negated=False)
-
-    return replace(
-        state,
-        phase=Phase.SELECT,
-        current_battle=None,
-        effect_queue=[],
-        last_restart_round=state.round_number,
-        pending_conditional_debuff_on_loss=(),
-        pending_draw_on_win=(),
-        pending_next_round_buff_on_win=(),
-        point_match_effects=(),
-        battle_janken_overrides=(),
-        win_condition_winner=None,
-    )
-
-
-@register("choose")
-def effect_choose(state: GameState, side: Side, card: Card) -> GameState:
-    choose_variant = {
-        "card_26": "yuriko_choose",
-        "c26": "yuriko_choose",
-        "card_45": "karen_choose",
-        "c45": "karen_choose",
-    }.get(card.id, "generic_choose")
-    ctx = PendingEffectContext(
-        side=side,
-        card_id=card.id,
-        effect="choose",
-        payload={"choose_variant": choose_variant},
-    )
-    state = replace(
-        state, phase=Phase.INTERACTIVE_EFFECT, effect_step=0, pending_effect_context=ctx
-    )
-    return replace(
-        state, battle_log=state.battle_log + [f"{card.name}の効果発動: 選択待機中..."]
-    )
-
-
-@register("choose_multiple")
-def effect_choose_multiple(state: GameState, side: Side, card: Card) -> GameState:
-    p_state = get_player_state(state, side)
-    if not p_state.hand and not p_state.deck:
-        return replace(
-            state,
-            battle_log=state.battle_log
-            + [f"{card.name}の効果発動: 発動可能な効果がないため不発"],
-        )
-
-    ctx = PendingEffectContext(side=side, card_id=card.id, effect="choose_multiple")
-    state = replace(
-        state, phase=Phase.INTERACTIVE_EFFECT, effect_step=0, pending_effect_context=ctx
-    )
-    return replace(
-        state,
-        battle_log=state.battle_log + [f"{card.name}の効果発動: 複数選択待機中..."],
-    )
-
-
-@register("copy_hand")
-def effect_copy_hand(state: GameState, side: Side, card: Card) -> GameState:
-    p_state = get_player_state(state, side)
-    if not p_state.hand:
-        return state
-
-    ctx = PendingEffectContext(side=side, card_id=card.id, effect="copy_hand")
-    state = replace(
-        state, phase=Phase.INTERACTIVE_EFFECT, effect_step=0, pending_effect_context=ctx
-    )
-    return replace(
-        state,
-        battle_log=state.battle_log + [f"{card.name}の効果発動: 対象選択待機中..."],
-    )
-
-
-@register("copy_effect")
-def effect_copy_effect(state: GameState, side: Side, card: Card) -> GameState:
-    p_state = get_player_state(state, side)
-    if not p_state.deck:
-        return state
-
-    top_card = p_state.deck[0]
-    new_deck = p_state.deck[1:] + [top_card]
-    state = update_player(state, side, deck=new_deck)
-
-    state = replace(
-        state,
-        battle_log=state.battle_log
-        + [f"{card.name}の効果発動: 山札の一番上({top_card.name})の効果を発動"],
-    )
-
-    if top_card.effect and top_card.effect.type.value == "copy_effect":
-        return replace(
-            state, battle_log=state.battle_log + ["-> copy_effectのため不発"]
-        )
-    if top_card.effect and top_card.effect.type.value in _registry:
-        new_queue = [(side, top_card)] + state.effect_queue
-        return replace(state, effect_queue=new_queue)
-
-    return state
-
-
-@register("search_and_swap")
-def effect_search_and_swap(state: GameState, side: Side, card: Card) -> GameState:
-    ctx = PendingEffectContext(side=side, card_id=card.id, effect="search_and_swap")
-    state = replace(
-        state, phase=Phase.INTERACTIVE_EFFECT, effect_step=0, pending_effect_context=ctx
-    )
-    return replace(
-        state,
-        battle_log=state.battle_log + [f"{card.name}の効果発動: 交換選択待機中..."],
-    )
-
-
-@register("swap")
-def effect_swap(state: GameState, side: Side, card: Card) -> GameState:
-    ctx = PendingEffectContext(side=side, card_id=card.id, effect="swap")
-    state = replace(
-        state, phase=Phase.INTERACTIVE_EFFECT, effect_step=0, pending_effect_context=ctx
-    )
-    return replace(
-        state,
-        battle_log=state.battle_log + [f"{card.name}の効果発動: 入れ替え選択待機中..."],
-    )
-
-
-@register("tutor_play")
-def effect_tutor_play(state: GameState, side: Side, card: Card) -> GameState:
-    ctx = PendingEffectContext(side=side, card_id=card.id, effect="tutor_play")
-    state = replace(
-        state, phase=Phase.INTERACTIVE_EFFECT, effect_step=0, pending_effect_context=ctx
-    )
-    return replace(
-        state,
-        battle_log=state.battle_log
-        + [f"{card.name}の効果発動: 場に出すカード選択待機中..."],
-    )
-
-
-@register("swap_opponent")
-def effect_swap_opponent(state: GameState, side: Side, card: Card) -> GameState:
-    if _opponent_is_immune(state, side):
-        return _immune_blocked_state(state, card)
-
-    ctx = PendingEffectContext(side=side, card_id=card.id, effect="swap_opponent")
-    state = replace(
-        state, phase=Phase.INTERACTIVE_EFFECT, effect_step=0, pending_effect_context=ctx
-    )
-    return replace(
-        state,
-        battle_log=state.battle_log
-        + [f"{card.name}の効果発動: 相手手札確認・入れ替え待機中..."],
-    )
-
-
-@register("removal")
-def effect_removal(state: GameState, side: Side, card: Card) -> GameState:
-    if _opponent_is_immune(state, side):
-        return _immune_blocked_state(state, card)
-
-    ctx = PendingEffectContext(side=side, card_id=card.id, effect="removal")
-    state = replace(
-        state, phase=Phase.INTERACTIVE_EFFECT, effect_step=0, pending_effect_context=ctx
-    )
-    return replace(
-        state, battle_log=state.battle_log + [f"{card.name}の効果発動: 発動待機中..."]
-    )
-
-
-@register("salvage")
-def effect_salvage(state: GameState, side: Side, card: Card) -> GameState:
-    ctx = PendingEffectContext(side=side, card_id=card.id, effect="salvage")
-    state = replace(
-        state, phase=Phase.INTERACTIVE_EFFECT, effect_step=0, pending_effect_context=ctx
-    )
-    return replace(
-        state,
-        battle_log=state.battle_log + [f"{card.name}の効果発動: 回収選択待機中..."],
-    )
-
-
-@register("reorder")
-def effect_reorder(state: GameState, side: Side, card: Card) -> GameState:
-    p_state = get_player_state(state, side)
-    if len(p_state.deck) <= 1:
-        return replace(
-            state,
-            battle_log=state.battle_log
-            + [f"{card.name}の効果発動: 並び替える山札がないため不発"],
-        )
-
-    ctx = PendingEffectContext(side=side, card_id=card.id, effect="reorder")
-    state = replace(
-        state, phase=Phase.INTERACTIVE_EFFECT, effect_step=0, pending_effect_context=ctx
-    )
-    return replace(
-        state,
-        battle_log=state.battle_log + [f"{card.name}の効果発動: 並び替え待機中..."],
-    )
-
-
-@register("curse")
-def effect_curse(state: GameState, side: Side, card: Card) -> GameState:
-    opp_side = get_opponent_side(side)
-    if _opponent_is_immune(state, side):
-        return _immune_blocked_state(state, card)
-
-    state = set_must_reveal_played_card(state, opp_side, True)
-    return replace(
-        state,
-        battle_log=state.battle_log
-        + [f"{card.name}の効果発動: 相手の次の出し札を公開"],
-    )
