+++ shadow_bout/effect_handlers/cards.py
@@ -1,0 +1,6 @@
+# Import modules for their @register side effects.
+from shadow_bout.effect_handlers import card_control as _card_control  # noqa: F401
+from shadow_bout.effect_handlers import card_draw as _card_draw  # noqa: F401
+from shadow_bout.effect_handlers import (
+    card_steal_copy as _card_steal_copy,  # noqa: F401
+)
