+++ shadow_bout/effect_handlers/points.py
@@ -1,0 +1,5 @@
+# Import modules for their @register side effects.
+from shadow_bout.effect_handlers import point_basic as _point_basic  # noqa: F401
+from shadow_bout.effect_handlers import (
+    point_conditional as _point_conditional,  # noqa: F401
+)
